+++ docs/release_notes.md
@@ -9,12 +9,14 @@ Additions
 - Add support for Warp backend task implementation
 - Add Warp-based RL examples: Cartpole, Ant, Humanoid
 - Add new Factory environments for place and screw: FactoryTaskNutBoltPlace and FactoryTaskNutBoltScrew
+- Add new camera-based Cartpole example: CartpoleCamera
+- Add new deformable environment showing Franka picking up a deformable tube: FrankaDeformable
 - Add support for running OIGE as an extension in Isaac Sim
 - Add options to filter collisions between environments and specify global collision filter paths to `RLTask.set_to_scene()`
 - Add multinode training support
 - Add dockerfile with OIGE
 - Add option to select kit app file from command line argument `kit_app`
-- Add new deformable environment showing Franka picking up a deformable tube: FrankaDeformable
+- Add `rendering_dt` parameter to the task config file for setting rendering dt. Defaults to the same value as the physics dt.
 
 Changes
 -------
@@ -27,7 +29,6 @@ Changes
 - The multi-threaded launch script `rlgames_train_mt.py` has been re-designed to support the extension workflow. This script can no longer be used to launch a training run from python. Please use `rlgames_train.py` instead.
 - Restructures for environments to support the new extension-based workflow
 - Add async workflow to factory pick environment to support extension-based workflow
-- The base RLTask has been moved to the Isaac Sim omni.isaa.gym module. Please make sure imports for RLTask are updated to the new module.
 - Update docker scripts with cache directories
 
 Fixes